Create Helium.Monitor.Setup.zip
Helium Monitor for Nebra (Windows)
v0.6.0.134

SSH is not working on this version ... it should put up a dialog to
try your miner's ip (press change). It goes into tray icon
(load's at startup) and the tray icon should change to red if your
miner has an error.

btw, it's big because it has chrome inside for the hellium explorer.

SSH monitoring is in works ....